src/trust-computation/requirements.txt
scikit-learn
pyyaml==5.4.1
numpy==1.20.03
flask==1.1.2
gunicorn==20.0.4
eventlet==0.30.2
requests==2.25.1
flask-restful==0.3.8